add examples as test cases
