+++ nucleus-examples/src/main.rs
@@ -1,3 +1,1 @@
-fn main() {
-    println!("Hello, world!");
-}
+pub fn main() {}
