+++ nucleus/cage/Cargo.toml
@@ -7,7 +7,7 @@ edition.workspace = true
 [dependencies]
 sp-runtime = { workspace = true }
 tokio = { version = "1", features = ["full"] }
-vrs-core-sdk = { path = "../core-sdk", features = ["host"] }
+vrs-core-sdk = { path = "../core-sdk" }
 vrs-primitives = { path = "../../primitives" }
 vrs-nucleus-runtime-api = { path = "../runtime-api" }
 codec = { workspace = true }
