+++ nucleus/core-macros/Cargo.toml
@@ -10,7 +10,7 @@ proc-macro = true
 [dependencies]
 syn = "2.0"
 quote = "1.0"
-vrs-core-sdk = { path = "../core-sdk" }
+vrs-core-sdk = { path = "../core-sdk", default-features = false }
 
 [features]
 default = ["std"]
